Reorganize app a bit
