--- README.md
@@ -1,39 +1,0 @@
-# [sqlalchemy-datatables](http://sqlalchemy-datatables.rtfd.org/)
-
-sqlalchemy-datatables is a library providing a SQLALchemy integration of jQuery DataTables. It helps you manage server side requests in your application.
-
-You can use it with [Pyramid](), [Flask]() or others that you can mention in a PR.
-
-It only depends on SQLAlchemy (obviously) :-).
-
-[![build status](https://img.shields.io/travis/Pegase745/sqlalchemy-datatables.svg?style=flat-square)](https://travis-ci.org/Pegase745/sqlalchemy-datatables/)
-[![PyPi version](https://img.shields.io/pypi/v/sqlalchemy-datatables.svg?style=flat-square)](https://pypi.python.org/pypi/sqlalchemy-datatables/)
-[![PyPi downloads](https://img.shields.io/pypi/dm/sqlalchemy-datatables.svg?style=flat-square)](https://pypi.python.org/pypi/sqlalchemy-datatables/)
-[![code quality](https://img.shields.io/scrutinizer/g/Pegase745/sqlalchemy-datatables.svg?style=flat-square)](https://scrutinizer-ci.com/g/Pegase745/sqlalchemy-datatables/)
-[![coveralls](https://img.shields.io/coveralls/Pegase745/sqlalchemy-datatables.svg?style=flat-square)](https://coveralls.io/r/Pegase745/sqlalchemy-datatables/)
-[![requirements](https://img.shields.io/requires/github/Pegase745/sqlalchemy-datatables.svg?style=flat-square)](https://requires.io/github/Pegase745/sqlalchemy-datatables/requirements/?branch=master/)
-
-### Installation
-
-To install the stable version:
-
-```
-pip install sqlalchemy-datatables
-```
-
-### Usage
-
-Coming soon
-
-### Documentation
-
-The latest documentation is available online on [Read the Docs](http://sqlalchemy-datatables.readthedocs.org/en/latest/).
-
-### Changelog
-
-This project adheres to [Semantic Versioning](http://semver.org/).
-Every release, is documented on the Github [releases](https://github.com/Pegase745/sqlalchemy-datatables/releases) page.
-
-### License
-
-The project is licensed under the MIT license.
