+++ README.rst
@@ -1,0 +1,57 @@
+=================================================================
+`sqlalchemy-datatables <http://sqlalchemy-datatables.rtfd.org/>`_
+=================================================================
+
+sqlalchemy-datatables is a library providing an SQLALchemy integration of jQuery DataTables. It helps you manage server side requests in your application.
+
+It is framework agnostic, tested with `Pyramid <http://>`_, `Flask <http://>`_ and `Django <http://>`_ mainly.
+
+It only depends on `SQLAlchemy <http://>`_.
+
+|Build Status| |PyPi Version| |PyPi Downloads| |Code Quality| |Coveralls|
+
+.. |Build Status| image:: https://img.shields.io/travis/Pegase745/sqlalchemy-datatables.svg?style=flat-square
+    :target: https://travis-ci.org/Pegase745/sqlalchemy-datatables
+
+.. |PyPi Version| image:: https://img.shields.io/pypi/v/sqlalchemy-datatables.svg?style=flat-square
+    :target: https://pypi.python.org/pypi/sqlalchemy-datatables/
+
+.. |PyPi Downloads| image:: https://img.shields.io/pypi/dm/sqlalchemy-datatables.svg?style=flat-square
+    :target: https://pypi.python.org/pypi/sqlalchemy-datatables/
+
+.. |Code Quality| image:: https://img.shields.io/scrutinizer/g/Pegase745/sqlalchemy-datatables.svg?style=flat-square
+    :target: https://scrutinizer-ci.com/g/Pegase745/sqlalchemy-datatables
+
+.. |Coveralls| image:: https://img.shields.io/coveralls/Pegase745/sqlalchemy-datatables.svg?style=flat-square
+    :target: https://coveralls.io/r/Pegase745/sqlalchemy-datatables
+
+Installation
+------------
+
+To install the stable version:
+
+.. code-block:: bash
+
+    pip install sqlalchemy-datatables
+
+
+Usage
+-----
+
+Coming soon
+
+Documentation
+-------------
+
+The latest documentation is available online on `Read the Docs <http://sqlalchemy-datatables.readthedocs.org/en/latest/>`_.
+
+Changelog
+---------
+
+This project adheres to `Semantic Versioning <http://semver.org/>`_.
+Every release, is documented on the Github `releases <https://github.com/Pegase745/sqlalchemy-datatables/releases>`_ page.
+
+License
+-------
+
+The project is licensed under the MIT license.
