+++ requirements.txt
@@ -1,3 +1,1 @@
-SQLAlchemy==1.0.9
-pyramid==1.5.7
-zope.sqlalchemy==0.7.6
+SQLAlchemy
