+++ setup.py
@@ -1,14 +1,5 @@
-#!/usr/bin/env python
-import sys
-
-from setuptools import setup, find_packages
-
-if sys.version_info >= (3, 0):
-    def my_open(path, mode):
-        return open(path, mode, newline='')
-else:
-    def my_open(path, mode):
-        return open(path, mode + 'b')
+"""App configuration."""
+from setuptools import setup
 
 
 __VERSION__ = '0.1.7'
@@ -24,7 +15,7 @@ setup(
     author='Michel Nemnom',
     author_email='michel.nemnom+pypi@gmail.com',
     url='https://github.com/pegase745/sqlalchemy-datatables',
-    packages=find_packages(exclude=['test_app']),
+    packages=['datatables'],
     include_package_data=True,
     zip_safe=False,
     install_requires=['sqlalchemy'],
